fix(bittensor): review fixes for validator subnet PR
- Extract shared ValidatorPicker components (ValidatorSortMethodButton,
  ValidatorRows, ValidatorRow, ValidatorRowSkeleton) to reduce duplication
  between ChangeValidatorSelect and BittensorValidatorPicker
- Add fallback to stakingPools.bittensor for default validator when new
  remote config fields are not yet populated
- Use ROOT_NETUID constant instead of hardcoded 0 in BittensorSubnetSelect
- Remove unnecessary useCallback wrappers around setSortMethod
- Fix double t() translation in sort method button labels
- Add test coverage for totalStaked, totalStakers, apr sort modes
- Add TODO noting composite sort metric normalization issue
