+++ apps/extension/src/ui/domains/Staking/Bond/hooks/useBondButton.ts
@@ -152,7 +152,8 @@ const getBondableBalance = (
    * Bittensor Staking
    */
   if (token?.type === "substrate-native" && bittensorNetworkIds.includes(token.networkId)) {
-    const defaultHotkey = remoteConfig.bittensor.defaultValidators[0] as string | undefined
+    const defaultHotkey = (remoteConfig.bittensor.defaultValidators[0] ??
+      remoteConfig.stakingPools.bittensor?.[0]) as string | undefined
 
     const isBonding = allBalances.each.some(
       (b) =>
