+++ apps/extension/src/ui/domains/TaoDashboard/subnet/swap/BittensorValidatorPicker.tsx
@@ -1,30 +1,20 @@
-import { type DotNetworkId, subNativeTokenId, type TokenId } from "@talismn/chaindata-provider"
-import { GlobeIcon, LockIcon, TalismanHandIcon, ToolbarSortIcon, UserIcon } from "@talismn/icons"
-import { cn, planckToTokens } from "@talismn/util"
-import { useVirtualizer } from "@tanstack/react-virtual"
-import {
-  ContextMenu,
-  ContextMenuContent,
-  ContextMenuOptionItem,
-  ContextMenuTrigger,
-} from "@ui/components/ContextMenu"
-import { ScrollContainer, useScrollContainer } from "@ui/components/ScrollContainer"
+import { type DotNetworkId, subNativeTokenId } from "@talismn/chaindata-provider"
+import { cn } from "@talismn/util"
+import { ScrollContainer } from "@ui/components/ScrollContainer"
 import { SearchInputControlled } from "@ui/components/SearchInputControlled"
-import { Tooltip, TooltipContent, TooltipTrigger } from "@ui/components/Tooltip"
-import { AccountIcon } from "@ui/domains/Account/AccountIcon"
-import { Address } from "@ui/domains/Account/Address"
-import { Tokens } from "@ui/domains/Asset/Tokens"
-import { EarnTypeBadge } from "@ui/domains/Earn/components/EarnTypeBadge"
+import {
+  ValidatorRowSkeleton,
+  ValidatorRows,
+  ValidatorSortMethodButton,
+} from "@ui/domains/Staking/Bittensor/components/ValidatorPicker"
 import {
   sortValidatorOptions,
   type ValidatorSortValue,
 } from "@ui/domains/Staking/Bittensor/utils/validatorSorting"
 import type { BondOption as BondOptionType } from "@ui/domains/Staking/hooks/bittensor/types"
 import { useCombinedBittensorValidatorsData } from "@ui/domains/Staking/hooks/bittensor/useCombinedBittensorValidatorsData"
-import { useToken } from "@ui/state/chaindata"
 import {
   type FC,
-  useCallback,
   useDeferredValue,
   useEffect,
   useMemo,
@@ -46,10 +36,6 @@ export const BittensorValidatorPicker: FC<{
   const [sortMethod, setSortMethod] = useState<ValidatorSortValue>("featured")
   const [rawSearch, setSearch] = useState<string>("")
   const search = useDeferredValue(rawSearch)
-
-  const handleSortMethodChange = useCallback((method: ValidatorSortValue) => {
-    setSortMethod(method)
-  }, [])
 
   const [sortedValidators, setSortedValidators] = useState<BondOptionType[] | undefined>(() =>
     combinedValidatorsData.length
@@ -103,7 +89,11 @@ export const BittensorValidatorPicker: FC<{
             autoFocus
           />
         </div>
-        <SortMethodButton method={sortMethod} onChange={handleSortMethodChange} />
+        <ValidatorSortMethodButton
+          method={sortMethod}
+          onChange={setSortMethod}
+          apyLabel={t("30 days APY")}
+        />
       </div>
       <div className="flex w-full grow flex-col gap-2 overflow-hidden">
         <div className="flex justify-between pr-12 pl-[6rem] text-body-disabled text-sm">
@@ -139,246 +129,5 @@ export const BittensorValidatorPicker: FC<{
         </ScrollContainer>
       </div>
     </div>
-  )
-}
-
-const SortMethodButton: FC<{
-  method: ValidatorSortValue
-  onChange: (method: ValidatorSortValue) => void
-}> = ({ method, onChange }) => {
-  const { t } = useTranslation()
-
-  const sortMethods = useMemo<{ label: string; value: ValidatorSortValue }[]>(
-    () => [
-      { label: t("Featured"), value: "featured" },
-      { label: t("Total Staked"), value: "totalStaked" },
-      { label: t("Name"), value: "name" },
-      { label: t("N° of Stakers"), value: "totalStakers" },
-      { label: t("30 days APY"), value: "apr" },
-    ],
-    [t]
-  )
-
-  const selected = useMemo(
-    () => sortMethods.find((sortMethod) => sortMethod.value === method),
-    [method, sortMethods]
-  )
-
-  return (
-    <ContextMenu placement="bottom-end">
-      <ContextMenuTrigger asChild>
-        <button
-          type="button"
-          className="flex h-full items-center gap-4 text-nowrap rounded-sm border border-grey-850 bg-field px-[8px] py-[6px] text-body-secondary text-sm hover:bg-grey-800 hover:text-grey-300"
-        >
-          <div>{selected?.label}</div>
-          <ToolbarSortIcon className="size-10" />
-        </button>
-      </ContextMenuTrigger>
-      <ContextMenuContent>
-        {sortMethods.map((sortMethod) => (
-          <ContextMenuOptionItem
-            key={sortMethod.value}
-            label={t(sortMethod.label)}
-            selected={sortMethod.value === method}
-            onClick={() => onChange(sortMethod.value)}
-          />
-        ))}
-      </ContextMenuContent>
-    </ContextMenu>
-  )
-}
-
-const ValidatorRows: FC<{
-  taoTokenId: string
-  validators: BondOptionType[]
-  selectedHotkey?: string | null
-  isLoading?: boolean
-  onSelect: (hotkey: string) => void
-}> = ({ taoTokenId, validators, selectedHotkey, isLoading, onSelect }) => {
-  const { ref: refContainer } = useScrollContainer()
-  const ref = useRef<HTMLDivElement>(null)
-
-  const virtualizer = useVirtualizer({
-    count: validators.length,
-    estimateSize: () => 58,
-    overscan: 5,
-    getScrollElement: () => refContainer.current,
-  })
-
-  if (!validators.length) return null
-
-  return (
-    <div ref={ref}>
-      <div
-        className="relative w-full"
-        style={{
-          height: `${virtualizer.getTotalSize()}px`,
-        }}
-      >
-        {virtualizer.getVirtualItems().map((item) => {
-          const validator = validators[item.index]
-          if (!validator) return null
-
-          return (
-            <div
-              key={item.key}
-              className="absolute top-0 left-0 w-full"
-              style={{
-                height: `${item.size}px`,
-                transform: `translateY(${item.start}px)`,
-              }}
-              data-testid="token-picker-row"
-            >
-              <ValidatorRow
-                key={item.key}
-                isSelected={validator.hotkey === selectedHotkey}
-                option={validator}
-                taoTokenId={taoTokenId}
-                onClick={() => onSelect(validator.hotkey)}
-                isLoading={isLoading}
-              />
-            </div>
-          )
-        })}
-      </div>
-    </div>
-  )
-}
-
-const ValidatorRowSkeleton = () => {
-  return (
-    <div className="flex h-[5.8rem] w-full shrink-0 items-center gap-6 px-12 pl-8 text-left">
-      <div className="size-16 animate-pulse rounded-full bg-grey-750"></div>
-      <div className="grow space-y-[5px]">
-        <div className={"flex w-full justify-between font-bold text-body text-sm"}>
-          <div>
-            <div className="inline-block h-7 w-56 animate-pulse rounded-xs bg-grey-750"></div>
-          </div>
-          <div>
-            <div className="inline-block h-7 w-20 animate-pulse rounded-xs bg-grey-750"></div>
-          </div>
-        </div>
-        <div className="flex w-full items-center justify-between gap-2 text-right font-light text-body-secondary text-xs">
-          <div>
-            <div className="inline-block h-6 w-40 animate-pulse rounded-xs bg-grey-800"></div>
-          </div>
-          <div className="grow text-right">
-            <div className="inline-block h-6 w-36 animate-pulse rounded-xs bg-grey-800"></div>
-          </div>
-        </div>
-      </div>
-    </div>
-  )
-}
-
-const ValidatorRow: FC<{
-  option: BondOptionType
-  taoTokenId: TokenId
-  isSelected: boolean
-  isLoading?: boolean
-  onClick: () => void
-}> = ({ option, isSelected, isLoading, taoTokenId, onClick }) => {
-  const { t } = useTranslation()
-  const tao = useToken(taoTokenId)
-
-  return (
-    <button
-      type="button"
-      key={option.hotkey}
-      onClick={onClick}
-      className={cn(
-        "flex h-[5.8rem] w-full shrink-0 items-center gap-6 overflow-hidden px-12 pl-8 text-left hover:bg-grey-750 focus:bg-grey-700",
-        "disabled:cursor-not-allowed disabled:opacity-50",
-        isSelected && "bg-grey-800 text-body-secondary"
-      )}
-    >
-      <AccountIcon address={option.hotkey} className="size-16 shrink-0 text-xl" />
-      <div className="flex h-full grow flex-col justify-center gap-2 overflow-hidden">
-        <div className="flex w-full justify-between text-body text-sm">
-          <div className={cn(option.isRecommended && "font-bold text-primary")}>
-            <div className="flex items-center gap-2">
-              {option.name ? (
-                <Tooltip>
-                  <TooltipTrigger asChild>
-                    <div>{option.name}</div>
-                  </TooltipTrigger>
-                  <TooltipContent>{option.hotkey}</TooltipContent>
-                </Tooltip>
-              ) : (
-                <Address startCharCount={8} endCharCount={8} address={option.hotkey} />
-              )}
-              {option.isFeatured && (
-                <EarnTypeBadge className="mx-0 inline-flex items-center gap-2 rounded border-none bg-primary/10 px-4 py-2 font-light text-primary text-xs normal-case">
-                  <TalismanHandIcon className="size-8" />
-                  {t("Featured")}
-                </EarnTypeBadge>
-              )}
-            </div>
-          </div>
-          <div className={cn(isLoading && "animate-pulse")}>
-            {option.validatorYield?.thirty_day_apy
-              ? `${(Number(option.validatorYield?.thirty_day_apy) * 100).toFixed(2)}%`
-              : t("N/A")}
-          </div>
-        </div>
-        <div
-          className={cn(
-            "flex w-full justify-between text-body-secondary text-xs",
-            isLoading && "animate-pulse"
-          )}
-        >
-          <div className="flex items-center gap-4">
-            <Tooltip>
-              <TooltipTrigger asChild>
-                <div className="flex items-center gap-2">
-                  <LockIcon />
-                  <Tokens
-                    amount={planckToTokens(option.totalStaked.toString(), tao?.decimals ?? 9)}
-                    symbol={tao?.symbol}
-                    noCountUp
-                    noTooltip
-                  />
-                </div>
-              </TooltipTrigger>
-              <TooltipContent>
-                <div className="flex flex-col gap-2">
-                  <div>{t("Total staked in this validator:")}</div>
-                  <div>
-                    {planckToTokens(option.totalStaked.toString(), tao?.decimals ?? 9)}{" "}
-                    {tao?.symbol}
-                  </div>
-                </div>
-              </TooltipContent>
-            </Tooltip>
-            <div className="inline-block size-2 rounded-full bg-body-disabled" />
-            <Tooltip>
-              <TooltipTrigger asChild>
-                <div className="flex items-center gap-2">
-                  <UserIcon />
-                  {option.totalStakers}
-                </div>
-              </TooltipTrigger>
-              <TooltipContent>
-                {t("{{count}} nominators", { count: option.totalStakers })}
-              </TooltipContent>
-            </Tooltip>
-            <div className="inline-block size-2 rounded-full bg-body-disabled" />
-            <Tooltip>
-              <TooltipTrigger asChild>
-                <div className="flex items-center gap-2">
-                  <GlobeIcon />
-                  {option.subnets}
-                </div>
-              </TooltipTrigger>
-              <TooltipContent>
-                {t("Validating {{count}} subnets", { count: option.subnets })}
-              </TooltipContent>
-            </Tooltip>
-          </div>
-          <div>{/* save slot for later */}</div>
-        </div>
-      </div>
-    </button>
   )
 }
